+++ packages/javascript/src/i18n/en-US.ts
@@ -82,7 +82,7 @@ const translations: I18nTranslations = {
   /* |                          User Profile                         | */
   /* |---------------------------------------------------------------| */
 
-  'user.profile.title': 'User Profile',
+  'user.profile.title': 'Profile',
   'user.profile.update.generic.error': 'An error occurred while updating your profile. Please try again.',
 
   /* |---------------------------------------------------------------| */