+++ packages/react/src/contexts/Asgardeo/AsgardeoContext.ts
@@ -17,7 +17,7 @@
  */
 
 import {Context, createContext} from 'react';
-import {HttpRequestConfig, HttpResponse, Organization} from '@asgardeo/browser';
+import {HttpRequestConfig, HttpResponse, IdToken, Organization, SignInOptions} from '@asgardeo/browser';
 import AsgardeoReactClient from '../../AsgardeoReactClient';
 
 /**
@@ -79,6 +79,24 @@ export type AsgardeoContextProps = {
      */
     requestAll: (requestConfigs?: HttpRequestConfig[]) => Promise<HttpResponse<any>[]>;
   };
+  /**
+   * Optional additional parameters to be sent in the sign-in request.
+   * This can include custom parameters that your authorization server supports.
+   * These parameters will be included in the authorization request sent to the server.
+   * If not provided, no additional parameters will be sent.
+   *
+   * @example
+   * signInOptions: { prompt: "login", fidp: "OrganizationSSO" }
+   */
+  signInOptions?: SignInOptions;
+  /**
+   * Function to retrieve the decoded ID token.
+   * This function decodes the ID token and returns its payload.
+   * It can be used to access user claims and other information contained in the ID token.
+   *
+   * @returns A promise that resolves to the decoded ID token payload.
+   */
+  getDecodedIdToken?: () => Promise<IdToken>;
 };
 
 /**
@@ -104,6 +122,8 @@ const AsgardeoContext: Context<AsgardeoContextProps | null> = createContext<null
     request: () => null,
     requestAll: () => null,
   },
+  signInOptions: {},
+  getDecodedIdToken: null,
 });
 
 AsgardeoContext.displayName = 'AsgardeoContext';