+++ packages/express/CHANGELOG.md
@@ -1,11 +1,45 @@
 # @asgardeo/express
 
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @asgardeo/node@0.0.14
+
+## 0.0.12
+
+### Patch Changes
+
+- [#119](https://github.com/asgardeo/javascript/pull/119)
+  [`ad71f09`](https://github.com/asgardeo/javascript/commit/ad71f09af3440b6e0b8d3aa1e93d0cbc941a1df3) Thanks
+  [@brionmario](https://github.com/brionmario)! - Update Sign In
+
+- Updated dependencies
+  [[`ad71f09`](https://github.com/asgardeo/javascript/commit/ad71f09af3440b6e0b8d3aa1e93d0cbc941a1df3)]:
+  - @asgardeo/node@0.0.13
+
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @asgardeo/node@0.0.12
+
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  [[`a175ecf`](https://github.com/asgardeo/javascript/commit/a175ecf37a89e7c03e3fd45bb92a684c18dc5cf3)]:
+  - @asgardeo/node@0.0.11
+
 ## 0.0.9
 
 ### Patch Changes
 
 - Updated dependencies
-  [[`7e8ea1c`](https://github.com/asgardeo/web-ui-sdks/commit/7e8ea1ca9219c1c95404933e8261b2abfbcad767)]:
+  [[`7e8ea1c`](https://github.com/asgardeo/javascript/commit/7e8ea1ca9219c1c95404933e8261b2abfbcad767)]:
   - @asgardeo/node@0.0.10
 
 ## 0.0.8
@@ -40,36 +74,36 @@
 
 ### Patch Changes
 
-- [#87](https://github.com/asgardeo/web-ui-sdks/pull/87)
-  [`c0b7ebd`](https://github.com/asgardeo/web-ui-sdks/commit/c0b7ebd71adb258d3df9fc336dfcb122e6ff6434) Thanks
+- [#87](https://github.com/asgardeo/javascript/pull/87)
+  [`c0b7ebd`](https://github.com/asgardeo/javascript/commit/c0b7ebd71adb258d3df9fc336dfcb122e6ff6434) Thanks
   [@brionmario](https://github.com/brionmario)! - Fix B2B components
 
 - Updated dependencies
-  [[`c0b7ebd`](https://github.com/asgardeo/web-ui-sdks/commit/c0b7ebd71adb258d3df9fc336dfcb122e6ff6434)]:
+  [[`c0b7ebd`](https://github.com/asgardeo/javascript/commit/c0b7ebd71adb258d3df9fc336dfcb122e6ff6434)]:
   - @asgardeo/node@0.0.5
 
 ## 0.0.3
 
 ### Patch Changes
 
-- [#85](https://github.com/asgardeo/web-ui-sdks/pull/85)
-  [`1066fef`](https://github.com/asgardeo/web-ui-sdks/commit/1066fefa855aa646b3026f6e682657e3980ca99b) Thanks
+- [#85](https://github.com/asgardeo/javascript/pull/85)
+  [`1066fef`](https://github.com/asgardeo/javascript/commit/1066fefa855aa646b3026f6e682657e3980ca99b) Thanks
   [@brionmario](https://github.com/brionmario)! - Fix issues with profile
 
 - Updated dependencies
-  [[`1066fef`](https://github.com/asgardeo/web-ui-sdks/commit/1066fefa855aa646b3026f6e682657e3980ca99b)]:
+  [[`1066fef`](https://github.com/asgardeo/javascript/commit/1066fefa855aa646b3026f6e682657e3980ca99b)]:
   - @asgardeo/node@0.0.4
 
 ## 0.0.2
 
 ### Patch Changes
 
-- [#83](https://github.com/asgardeo/web-ui-sdks/pull/83)
-  [`9cebe25`](https://github.com/asgardeo/web-ui-sdks/commit/9cebe25b74c6429794ee583cd7f110f0a951851f) Thanks
+- [#83](https://github.com/asgardeo/javascript/pull/83)
+  [`9cebe25`](https://github.com/asgardeo/javascript/commit/9cebe25b74c6429794ee583cd7f110f0a951851f) Thanks
   [@brionmario](https://github.com/brionmario)! - Fix NPEs
 
 - Updated dependencies
-  [[`9cebe25`](https://github.com/asgardeo/web-ui-sdks/commit/9cebe25b74c6429794ee583cd7f110f0a951851f)]:
+  [[`9cebe25`](https://github.com/asgardeo/javascript/commit/9cebe25b74c6429794ee583cd7f110f0a951851f)]:
   - @asgardeo/node@0.0.3
 
 ## 0.0.1