+++ packages/react-router/CHANGELOG.md
@@ -1,11 +1,48 @@
 # @asgardeo/react-router
 
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  [[`f01f421`](https://github.com/asgardeo/javascript/commit/f01f421acb6bafe10142cb329641749cc4cd64bb)]:
+  - @asgardeo/react@0.5.16
+
+## 0.0.8
+
+### Patch Changes
+
+- [#119](https://github.com/asgardeo/javascript/pull/119)
+  [`ad71f09`](https://github.com/asgardeo/javascript/commit/ad71f09af3440b6e0b8d3aa1e93d0cbc941a1df3) Thanks
+  [@brionmario](https://github.com/brionmario)! - Update Sign In
+
+- Updated dependencies
+  [[`ad71f09`](https://github.com/asgardeo/javascript/commit/ad71f09af3440b6e0b8d3aa1e93d0cbc941a1df3)]:
+  - @asgardeo/react@0.5.15
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  [[`6ba5b4b`](https://github.com/asgardeo/javascript/commit/6ba5b4bed7e6d64fbedb7d9099ce8126c58de7eb)]:
+  - @asgardeo/react@0.5.14
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  [[`a1a6afe`](https://github.com/asgardeo/javascript/commit/a1a6afe776a527d7485920f3ee448131196300aa),
+  [`a175ecf`](https://github.com/asgardeo/javascript/commit/a175ecf37a89e7c03e3fd45bb92a684c18dc5cf3)]:
+  - @asgardeo/react@0.5.13
+
 ## 0.0.5
 
 ### Patch Changes
 
 - Updated dependencies
-  [[`7e8ea1c`](https://github.com/asgardeo/web-ui-sdks/commit/7e8ea1ca9219c1c95404933e8261b2abfbcad767)]:
+  [[`7e8ea1c`](https://github.com/asgardeo/javascript/commit/7e8ea1ca9219c1c95404933e8261b2abfbcad767)]:
   - @asgardeo/react@0.5.12
 
 ## 0.0.4
@@ -20,7 +57,7 @@
 ### Patch Changes
 
 - Updated dependencies
-  [[`399a2e1`](https://github.com/asgardeo/web-ui-sdks/commit/399a2e14473c545a3c549aa626a028f1640f227f)]:
+  [[`399a2e1`](https://github.com/asgardeo/javascript/commit/399a2e14473c545a3c549aa626a028f1640f227f)]:
   - @asgardeo/react@0.5.10
 
 ## 0.0.2
@@ -28,21 +65,21 @@
 ### Patch Changes
 
 - Updated dependencies
-  [[`b50ea2b`](https://github.com/asgardeo/web-ui-sdks/commit/b50ea2b5be3f8a8056fe64edd2b09cba8ee69691)]:
+  [[`b50ea2b`](https://github.com/asgardeo/javascript/commit/b50ea2b5be3f8a8056fe64edd2b09cba8ee69691)]:
   - @asgardeo/react@0.5.9
 
 ## 0.0.1
 
 ### Patch Changes
 
-- [#94](https://github.com/asgardeo/web-ui-sdks/pull/94)
-  [`cb918a3`](https://github.com/asgardeo/web-ui-sdks/commit/cb918a30a4c195f0ca06f672d6146bbe4d555f27) Thanks
+- [#94](https://github.com/asgardeo/javascript/pull/94)
+  [`cb918a3`](https://github.com/asgardeo/javascript/commit/cb918a30a4c195f0ca06f672d6146bbe4d555f27) Thanks
   [@brionmario](https://github.com/brionmario)! - This update addresses issues in the `@asgardeo/react-router` package
   and exposes the `signInSilently` method from the `@asgardeo/react` package. The changes ensure that the
   `ProtectedRoute` component works correctly with the new sign-in functionality.
 
 - Updated dependencies
-  [[`cb918a3`](https://github.com/asgardeo/web-ui-sdks/commit/cb918a30a4c195f0ca06f672d6146bbe4d555f27)]:
+  [[`cb918a3`](https://github.com/asgardeo/javascript/commit/cb918a30a4c195f0ca06f672d6146bbe4d555f27)]:
   - @asgardeo/react@0.5.8
 
 ## 1.0.0
@@ -50,5 +87,5 @@
 ### Patch Changes
 
 - Updated dependencies
-  [[`3aa86ce`](https://github.com/asgardeo/web-ui-sdks/commit/3aa86cebb7077f516994a9a6773f3389e7d6a6c8)]:
+  [[`3aa86ce`](https://github.com/asgardeo/javascript/commit/3aa86cebb7077f516994a9a6773f3389e7d6a6c8)]:
   - @asgardeo/react@0.5.0