+++ samples/__legacy__/nuxt-vite/CHANGELOG.md
@@ -5,5 +5,5 @@
 ### Patch Changes
 
 - Updated dependencies
-  [[`74afcc5`](https://github.com/asgardeo/web-ui-sdks/commit/74afcc5bbf3dcfd8a2ec0c0026b709eafbe609a1)]:
+  [[`74afcc5`](https://github.com/asgardeo/javascript/commit/74afcc5bbf3dcfd8a2ec0c0026b709eafbe609a1)]:
   - @asgardeo/nuxt@0.0.1